Activated support for some additional archive formats (e.g. tar.gz).

+++ src/net/sourceforge/docfetcher/model/TreeIndex.java
@@ -135,7 +135,7 @@ public abstract class TreeIndex <
 	}
 	
 	private void clear(boolean removeTopLevel) {
-		if (fileIndexDir != null) {
+		if (fileIndexDir != null && fileIndexDir.exists()) {
 			try {
 				if (removeTopLevel)
 					Files.deleteRecursively(fileIndexDir);
